Move filters utilities to a separate file
Breaking things into smaller files will make things
easier to manage
